Write log using t.Log in registry unit tests
