+++ pkg/dockerregistry/server/pullthroughblobstore_test.go
@@ -26,6 +26,9 @@ import (
 )
 
 func TestPullthroughServeBlob(t *testing.T) {
+	backgroundCtx := context.Background()
+	backgroundCtx = registrytest.WithTestLogger(backgroundCtx, t)
+
 	namespace, name := "user", "app"
 	repoName := fmt.Sprintf("%s/%s", namespace, name)
 	installFakeAccessController(t)
@@ -36,7 +39,7 @@ func TestPullthroughServeBlob(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	remoteRegistryServer := createTestRegistryServer(t, context.Background())
+	remoteRegistryServer := createTestRegistryServer(t, backgroundCtx)
 	defer remoteRegistryServer.Close()
 
 	serverURL, err := url.Parse(remoteRegistryServer.URL)
@@ -46,7 +49,7 @@ func TestPullthroughServeBlob(t *testing.T) {
 	os.Setenv("OPENSHIFT_DEFAULT_REGISTRY", serverURL.Host)
 	testImage.DockerImageReference = fmt.Sprintf("%s/%s@%s", serverURL.Host, repoName, testImage.Name)
 
-	fos, imageClient := registrytest.NewFakeOpenShiftWithClient()
+	fos, imageClient := registrytest.NewFakeOpenShiftWithClient(backgroundCtx)
 	registrytest.AddImageStream(t, fos, namespace, name, map[string]string{
 		imageapi.InsecureRepositoryAnnotation: "true",
 	})
@@ -135,8 +138,8 @@ func TestPullthroughServeBlob(t *testing.T) {
 	} {
 		localBlobStore := newTestBlobStore(nil, tc.localBlobs)
 
-		ctx := WithTestPassthroughToUpstream(context.Background(), false)
-		repo := newTestRepository(t, namespace, name, testRepositoryOptions{
+		ctx := WithTestPassthroughToUpstream(backgroundCtx, false)
+		repo := newTestRepository(ctx, t, namespace, name, testRepositoryOptions{
 			client:            registryclient.NewFakeRegistryAPIClient(nil, imageClient),
 			enablePullThrough: true,
 		})
@@ -278,7 +281,9 @@ func TestPullthroughServeNotSeekableBlob(t *testing.T) {
 	}
 	testImage.DockerImageReference = fmt.Sprintf("%s/%s@%s", serverURL.Host, repoName, testImage.Name)
 
-	fos, imageClient := registrytest.NewFakeOpenShiftWithClient()
+	ctx := context.Background()
+	ctx = registrytest.WithTestLogger(ctx, t)
+	fos, imageClient := registrytest.NewFakeOpenShiftWithClient(ctx)
 	registrytest.AddImageStream(t, fos, namespace, name, map[string]string{
 		imageapi.InsecureRepositoryAnnotation: "true",
 	})
@@ -286,8 +291,8 @@ func TestPullthroughServeNotSeekableBlob(t *testing.T) {
 
 	localBlobStore := newTestBlobStore(nil, nil)
 
-	ctx := WithTestPassthroughToUpstream(context.Background(), false)
-	repo := newTestRepository(t, namespace, name, testRepositoryOptions{
+	ctx = WithTestPassthroughToUpstream(ctx, false)
+	repo := newTestRepository(ctx, t, namespace, name, testRepositoryOptions{
 		client:            registryclient.NewFakeRegistryAPIClient(nil, imageClient),
 		enablePullThrough: true,
 	})
@@ -361,7 +366,10 @@ func TestPullthroughServeBlobInsecure(t *testing.T) {
 	installFakeAccessController(t)
 	setPassthroughBlobDescriptorServiceFactory()
 
-	remoteRegistryServer := createTestRegistryServer(t, context.Background())
+	backgroundCtx := context.Background()
+	backgroundCtx = registrytest.WithTestLogger(backgroundCtx, t)
+
+	remoteRegistryServer := createTestRegistryServer(t, backgroundCtx)
 	defer remoteRegistryServer.Close()
 
 	serverURL, err := url.Parse(remoteRegistryServer.URL)
@@ -598,77 +606,82 @@ func TestPullthroughServeBlobInsecure(t *testing.T) {
 			expectedBytesServed:   int64(m1img.DockerImageLayers[0].LayerSize),
 		},
 	} {
-		fos, imageClient := registrytest.NewFakeOpenShiftWithClient()
+		tc := tc
+		t.Run(tc.name, func(t *testing.T) {
+			ctx := registrytest.WithTestLogger(backgroundCtx, t)
 
-		tc.fakeOpenShiftInit(fos)
+			fos, imageClient := registrytest.NewFakeOpenShiftWithClient(ctx)
 
-		localBlobStore := newTestBlobStore(nil, tc.localBlobs)
+			tc.fakeOpenShiftInit(fos)
 
-		ctx := WithTestPassthroughToUpstream(context.Background(), false)
+			localBlobStore := newTestBlobStore(nil, tc.localBlobs)
 
-		repo := newTestRepository(t, namespace, repo1, testRepositoryOptions{
-			client:            registryclient.NewFakeRegistryAPIClient(nil, imageClient),
-			enablePullThrough: true,
+			ctx = WithTestPassthroughToUpstream(ctx, false)
+
+			repo := newTestRepository(ctx, t, namespace, repo1, testRepositoryOptions{
+				client:            registryclient.NewFakeRegistryAPIClient(nil, imageClient),
+				enablePullThrough: true,
+			})
+
+			ptbs := &pullthroughBlobStore{
+				BlobStore: localBlobStore,
+				repo:      repo,
+			}
+
+			req, err := http.NewRequest(tc.method, fmt.Sprintf("http://example.org/v2/user/app/blobs/%s", tc.blobDigest), nil)
+			if err != nil {
+				t.Fatalf("[%s] failed to create http request: %v", tc.name, err)
+			}
+			w := httptest.NewRecorder()
+
+			dgst := digest.Digest(tc.blobDigest)
+
+			_, err = ptbs.Stat(ctx, dgst)
+			if err != tc.expectedStatError {
+				t.Fatalf("[%s] Stat returned unexpected error: %#+v != %#+v", tc.name, err, tc.expectedStatError)
+			}
+			if err != nil || tc.expectedStatError != nil {
+				return
+			}
+			err = ptbs.ServeBlob(ctx, w, req, dgst)
+			if err != nil {
+				t.Errorf("[%s] unexpected ServeBlob error: %v", tc.name, err)
+				return
+			}
+
+			clstr := w.Header().Get("Content-Length")
+			if cl, err := strconv.ParseInt(clstr, 10, 64); err != nil {
+				t.Errorf(`[%s] unexpected Content-Length: %q != "%d"`, tc.name, clstr, tc.expectedContentLength)
+			} else {
+				if cl != tc.expectedContentLength {
+					t.Errorf("[%s] Content-Length does not match expected size: %d != %d", tc.name, cl, tc.expectedContentLength)
+				}
+			}
+			if w.Header().Get("Content-Type") != "application/octet-stream" {
+				t.Errorf("[%s] Content-Type does not match expected: %q != %q", tc.name, w.Header().Get("Content-Type"), "application/octet-stream")
+			}
+
+			body := w.Body.Bytes()
+			if int64(len(body)) != tc.expectedBytesServed {
+				t.Errorf("[%s] unexpected size of body: %d != %d", tc.name, len(body), tc.expectedBytesServed)
+			}
+
+			for name, expCount := range tc.expectedLocalCalls {
+				count := localBlobStore.calls[name]
+				if count != expCount {
+					t.Errorf("[%s] expected %d calls to method %s of local blob store, not %d", tc.name, expCount, name, count)
+				}
+			}
+			for name, count := range localBlobStore.calls {
+				if _, exists := tc.expectedLocalCalls[name]; !exists {
+					t.Errorf("[%s] expected no calls to method %s of local blob store, got %d", tc.name, name, count)
+				}
+			}
+
+			if localBlobStore.bytesServed != tc.expectedBytesServedLocally {
+				t.Errorf("[%s] unexpected number of bytes served locally: %d != %d", tc.name, localBlobStore.bytesServed, tc.expectedBytesServed)
+			}
 		})
-
-		ptbs := &pullthroughBlobStore{
-			BlobStore: localBlobStore,
-			repo:      repo,
-		}
-
-		req, err := http.NewRequest(tc.method, fmt.Sprintf("http://example.org/v2/user/app/blobs/%s", tc.blobDigest), nil)
-		if err != nil {
-			t.Fatalf("[%s] failed to create http request: %v", tc.name, err)
-		}
-		w := httptest.NewRecorder()
-
-		dgst := digest.Digest(tc.blobDigest)
-
-		_, err = ptbs.Stat(ctx, dgst)
-		if err != tc.expectedStatError {
-			t.Fatalf("[%s] Stat returned unexpected error: %#+v != %#+v", tc.name, err, tc.expectedStatError)
-		}
-		if err != nil || tc.expectedStatError != nil {
-			continue
-		}
-		err = ptbs.ServeBlob(ctx, w, req, dgst)
-		if err != nil {
-			t.Errorf("[%s] unexpected ServeBlob error: %v", tc.name, err)
-			continue
-		}
-
-		clstr := w.Header().Get("Content-Length")
-		if cl, err := strconv.ParseInt(clstr, 10, 64); err != nil {
-			t.Errorf(`[%s] unexpected Content-Length: %q != "%d"`, tc.name, clstr, tc.expectedContentLength)
-		} else {
-			if cl != tc.expectedContentLength {
-				t.Errorf("[%s] Content-Length does not match expected size: %d != %d", tc.name, cl, tc.expectedContentLength)
-			}
-		}
-		if w.Header().Get("Content-Type") != "application/octet-stream" {
-			t.Errorf("[%s] Content-Type does not match expected: %q != %q", tc.name, w.Header().Get("Content-Type"), "application/octet-stream")
-		}
-
-		body := w.Body.Bytes()
-		if int64(len(body)) != tc.expectedBytesServed {
-			t.Errorf("[%s] unexpected size of body: %d != %d", tc.name, len(body), tc.expectedBytesServed)
-		}
-
-		for name, expCount := range tc.expectedLocalCalls {
-			count := localBlobStore.calls[name]
-			if count != expCount {
-				t.Errorf("[%s] expected %d calls to method %s of local blob store, not %d", tc.name, expCount, name, count)
-			}
-		}
-		for name, count := range localBlobStore.calls {
-			if _, exists := tc.expectedLocalCalls[name]; !exists {
-				t.Errorf("[%s] expected no calls to method %s of local blob store, got %d", tc.name, name, count)
-			}
-		}
-
-		if localBlobStore.bytesServed != tc.expectedBytesServedLocally {
-			t.Errorf("[%s] unexpected number of bytes served locally: %d != %d", tc.name, localBlobStore.bytesServed, tc.expectedBytesServed)
-		}
 	}
 }
