+++ pkg/dockerregistry/server/tagservice_test.go
@@ -18,7 +18,10 @@ func TestTagGet(t *testing.T) {
 	repo := "app"
 	tag := "latest"
 
-	fos, imageClient := registrytest.NewFakeOpenShiftWithClient()
+	backgroundCtx := context.Background()
+	backgroundCtx = registrytest.WithTestLogger(backgroundCtx, t)
+
+	fos, imageClient := registrytest.NewFakeOpenShiftWithClient(backgroundCtx)
 	testImage := registrytest.AddRandomImage(t, fos, namespace, repo, tag)
 
 	testcases := []struct {
@@ -68,7 +71,7 @@ func TestTagGet(t *testing.T) {
 			testImage.Annotations[imageapi.ManagedByOpenShiftAnnotation] = "false"
 		}
 
-		r := newTestRepository(t, namespace, repo, testRepositoryOptions{
+		r := newTestRepository(backgroundCtx, t, namespace, repo, testRepositoryOptions{
 			client:            registryclient.NewFakeRegistryAPIClient(nil, imageClient),
 			enablePullThrough: tc.pullthrough,
 		})
@@ -78,7 +81,7 @@ func TestTagGet(t *testing.T) {
 			repo:       r,
 		}
 
-		resultDesc, err := ts.Get(context.Background(), tc.tagName)
+		resultDesc, err := ts.Get(backgroundCtx, tc.tagName)
 
 		switch err.(type) {
 		case distribution.ErrTagUnknown:
@@ -107,9 +110,12 @@ func TestTagGetWithoutImageStream(t *testing.T) {
 	repo := "app"
 	tag := "latest"
 
-	_, imageClient := registrytest.NewFakeOpenShiftWithClient()
+	ctx := context.Background()
+	ctx = registrytest.WithTestLogger(ctx, t)
 
-	r := newTestRepository(t, namespace, repo, testRepositoryOptions{
+	_, imageClient := registrytest.NewFakeOpenShiftWithClient(ctx)
+
+	r := newTestRepository(ctx, t, namespace, repo, testRepositoryOptions{
 		client: registryclient.NewFakeRegistryAPIClient(nil, imageClient),
 	})
 
@@ -118,7 +124,7 @@ func TestTagGetWithoutImageStream(t *testing.T) {
 		repo:       r,
 	}
 
-	_, err := ts.Get(context.Background(), tag)
+	_, err := ts.Get(ctx, tag)
 	if err == nil {
 		t.Fatalf("error expected")
 	}
@@ -134,7 +140,10 @@ func TestTagCreation(t *testing.T) {
 	repo := "app"
 	tag := "latest"
 
-	fos, imageClient := registrytest.NewFakeOpenShiftWithClient()
+	ctx := context.Background()
+	ctx = registrytest.WithTestLogger(ctx, t)
+
+	fos, imageClient := registrytest.NewFakeOpenShiftWithClient(ctx)
 	testImage := registrytest.AddRandomImage(t, fos, namespace, repo, tag)
 
 	testcases := []struct {
@@ -173,7 +182,7 @@ func TestTagCreation(t *testing.T) {
 			testImage.Annotations[imageapi.ManagedByOpenShiftAnnotation] = "false"
 		}
 
-		r := newTestRepository(t, namespace, repo, testRepositoryOptions{
+		r := newTestRepository(ctx, t, namespace, repo, testRepositoryOptions{
 			client:            registryclient.NewFakeRegistryAPIClient(nil, imageClient),
 			enablePullThrough: tc.pullthrough,
 		})
@@ -183,7 +192,7 @@ func TestTagCreation(t *testing.T) {
 			repo:       r,
 		}
 
-		err := ts.Tag(context.Background(), tc.tagName, tc.tagValue)
+		err := ts.Tag(ctx, tc.tagName, tc.tagValue)
 		if tc.expectedError {
 			if err == nil {
 				t.Fatalf("[%s] error expected", tc.title)
@@ -191,7 +200,7 @@ func TestTagCreation(t *testing.T) {
 			continue
 		}
 
-		r = newTestRepository(t, namespace, repo, testRepositoryOptions{
+		r = newTestRepository(ctx, t, namespace, repo, testRepositoryOptions{
 			client:            registryclient.NewFakeRegistryAPIClient(nil, imageClient),
 			enablePullThrough: tc.pullthrough,
 		})
@@ -201,7 +210,7 @@ func TestTagCreation(t *testing.T) {
 			repo:       r,
 		}
 
-		tag, err := ts.Get(context.Background(), tc.tagName)
+		tag, err := ts.Get(ctx, tc.tagName)
 		if err != nil {
 			t.Fatalf("[%s] unexpected error: %s", tc.title, err)
 		}
@@ -217,10 +226,13 @@ func TestTagCreationWithoutImageStream(t *testing.T) {
 	repo := "app"
 	tag := "latest"
 
-	fos, imageClient := registrytest.NewFakeOpenShiftWithClient()
+	ctx := context.Background()
+	ctx = registrytest.WithTestLogger(ctx, t)
+
+	fos, imageClient := registrytest.NewFakeOpenShiftWithClient(ctx)
 	anotherImage := registrytest.AddRandomImage(t, fos, namespace, repo+"-another", tag)
 
-	r := newTestRepository(t, namespace, repo, testRepositoryOptions{
+	r := newTestRepository(ctx, t, namespace, repo, testRepositoryOptions{
 		client: registryclient.NewFakeRegistryAPIClient(nil, imageClient),
 	})
 
@@ -229,7 +241,7 @@ func TestTagCreationWithoutImageStream(t *testing.T) {
 		repo:       r,
 	}
 
-	err := ts.Tag(context.Background(), tag, distribution.Descriptor{
+	err := ts.Tag(ctx, tag, distribution.Descriptor{
 		Digest: digest.Digest(anotherImage.Name),
 	})
 	if err == nil {
@@ -247,7 +259,10 @@ func TestTagDeletion(t *testing.T) {
 	repo := "app"
 	tag := "latest"
 
-	fos, imageClient := registrytest.NewFakeOpenShiftWithClient()
+	ctx := context.Background()
+	ctx = registrytest.WithTestLogger(ctx, t)
+
+	fos, imageClient := registrytest.NewFakeOpenShiftWithClient(ctx)
 	testImage := registrytest.AddRandomImage(t, fos, namespace, repo, tag)
 
 	testcases := []struct {
@@ -294,7 +309,7 @@ func TestTagDeletion(t *testing.T) {
 			testImage.Annotations[imageapi.ManagedByOpenShiftAnnotation] = "false"
 		}
 
-		r := newTestRepository(t, namespace, repo, testRepositoryOptions{
+		r := newTestRepository(ctx, t, namespace, repo, testRepositoryOptions{
 			client:            registryclient.NewFakeRegistryAPIClient(nil, imageClient),
 			enablePullThrough: tc.pullthrough,
 		})
@@ -304,7 +319,7 @@ func TestTagDeletion(t *testing.T) {
 			repo:       r,
 		}
 
-		err := ts.Untag(context.Background(), tc.tagName)
+		err := ts.Untag(ctx, tc.tagName)
 
 		switch err.(type) {
 		case distribution.ErrTagUnknown:
@@ -329,9 +344,12 @@ func TestTagDeletionWithoutImageStream(t *testing.T) {
 	repo := "app"
 	tag := "latest"
 
-	_, imageClient := registrytest.NewFakeOpenShiftWithClient()
+	ctx := context.Background()
+	ctx = registrytest.WithTestLogger(ctx, t)
 
-	r := newTestRepository(t, namespace, repo, testRepositoryOptions{
+	_, imageClient := registrytest.NewFakeOpenShiftWithClient(ctx)
+
+	r := newTestRepository(ctx, t, namespace, repo, testRepositoryOptions{
 		client: registryclient.NewFakeRegistryAPIClient(nil, imageClient),
 	})
 
@@ -340,7 +358,7 @@ func TestTagDeletionWithoutImageStream(t *testing.T) {
 		repo:       r,
 	}
 
-	err := ts.Untag(context.Background(), tag)
+	err := ts.Untag(ctx, tag)
 	if err == nil {
 		t.Fatalf("error expected")
 	}
@@ -356,7 +374,10 @@ func TestTagGetAll(t *testing.T) {
 	repo := "app"
 	tag := "latest"
 
-	fos, imageClient := registrytest.NewFakeOpenShiftWithClient()
+	ctx := context.Background()
+	ctx = registrytest.WithTestLogger(ctx, t)
+
+	fos, imageClient := registrytest.NewFakeOpenShiftWithClient(ctx)
 	testImage := registrytest.AddRandomImage(t, fos, namespace, repo, tag)
 
 	testcases := []struct {
@@ -390,7 +411,7 @@ func TestTagGetAll(t *testing.T) {
 			testImage.Annotations[imageapi.ManagedByOpenShiftAnnotation] = "false"
 		}
 
-		r := newTestRepository(t, namespace, repo, testRepositoryOptions{
+		r := newTestRepository(ctx, t, namespace, repo, testRepositoryOptions{
 			client:            registryclient.NewFakeRegistryAPIClient(nil, imageClient),
 			enablePullThrough: tc.pullthrough,
 		})
@@ -400,7 +421,7 @@ func TestTagGetAll(t *testing.T) {
 			repo:       r,
 		}
 
-		result, err := ts.All(context.Background())
+		result, err := ts.All(ctx)
 
 		if err != nil && !tc.expectedError {
 			t.Fatalf("[%s] unexpected error: %#+v", tc.title, err)
@@ -416,9 +437,12 @@ func TestTagGetAllWithoutImageStream(t *testing.T) {
 	namespace := "user"
 	repo := "app"
 
-	_, imageClient := registrytest.NewFakeOpenShiftWithClient()
+	ctx := context.Background()
+	ctx = registrytest.WithTestLogger(ctx, t)
 
-	r := newTestRepository(t, namespace, repo, testRepositoryOptions{
+	_, imageClient := registrytest.NewFakeOpenShiftWithClient(ctx)
+
+	r := newTestRepository(ctx, t, namespace, repo, testRepositoryOptions{
 		client: registryclient.NewFakeRegistryAPIClient(nil, imageClient),
 	})
 
@@ -427,7 +451,7 @@ func TestTagGetAllWithoutImageStream(t *testing.T) {
 		repo:       r,
 	}
 
-	_, err := ts.All(context.Background())
+	_, err := ts.All(ctx)
 	if err == nil {
 		t.Fatalf("error expected")
 	}
@@ -443,7 +467,10 @@ func TestTagLookup(t *testing.T) {
 	repo := "app"
 	tag := "latest"
 
-	fos, imageClient := registrytest.NewFakeOpenShiftWithClient()
+	ctx := context.Background()
+	ctx = registrytest.WithTestLogger(ctx, t)
+
+	fos, imageClient := registrytest.NewFakeOpenShiftWithClient(ctx)
 	testImage := registrytest.AddRandomImage(t, fos, namespace, repo, tag)
 
 	testcases := []struct {
@@ -488,7 +515,7 @@ func TestTagLookup(t *testing.T) {
 			testImage.Annotations[imageapi.ManagedByOpenShiftAnnotation] = "false"
 		}
 
-		r := newTestRepository(t, namespace, repo, testRepositoryOptions{
+		r := newTestRepository(ctx, t, namespace, repo, testRepositoryOptions{
 			client:            registryclient.NewFakeRegistryAPIClient(nil, imageClient),
 			enablePullThrough: tc.pullthrough,
 		})
@@ -498,7 +525,7 @@ func TestTagLookup(t *testing.T) {
 			repo:       r,
 		}
 
-		result, err := ts.Lookup(context.Background(), tc.tagValue)
+		result, err := ts.Lookup(ctx, tc.tagValue)
 
 		if err != nil {
 			if !tc.expectedError {
@@ -522,10 +549,13 @@ func TestTagLookupWithoutImageStream(t *testing.T) {
 	repo := "app"
 	tag := "latest"
 
-	fos, imageClient := registrytest.NewFakeOpenShiftWithClient()
+	ctx := context.Background()
+	ctx = registrytest.WithTestLogger(ctx, t)
+
+	fos, imageClient := registrytest.NewFakeOpenShiftWithClient(ctx)
 	anotherImage := registrytest.AddRandomImage(t, fos, namespace, repo+"-another", tag)
 
-	r := newTestRepository(t, namespace, repo, testRepositoryOptions{
+	r := newTestRepository(ctx, t, namespace, repo, testRepositoryOptions{
 		client: registryclient.NewFakeRegistryAPIClient(nil, imageClient),
 	})
 
@@ -534,7 +564,7 @@ func TestTagLookupWithoutImageStream(t *testing.T) {
 		repo:       r,
 	}
 
-	_, err := ts.Lookup(context.Background(), distribution.Descriptor{
+	_, err := ts.Lookup(ctx, distribution.Descriptor{
 		Digest: digest.Digest(anotherImage.Name),
 	})
 	if err == nil {
